Align contract term and chart labels
Fix contract term display and sorting across modules; store 6-month term correctly as 0.5; update sorting to use codes descending for trainees, partners, unions, and orders; ensure contract_term saved as float; improve chart labels visibility.

X-Lovable-Edit-ID: edt-3b766c62-69ae-40a6-a09b-45164cb40341

+++ src/pages/TraineeForm.tsx
@@ -581,7 +581,7 @@ function TraineeFormContent({ isEditMode, traineeId }: TraineeFormContentProps) 
         (traineeData as any).expected_entry_month = projectData.expected_entry_month;
       }
       if (projectData.contract_term) {
-        (traineeData as any).contract_term = parseInt(projectData.contract_term);
+        (traineeData as any).contract_term = parseFloat(projectData.contract_term);
       }
       if (projectData.interview_date) {
         (traineeData as any).interview_pass_date = projectData.interview_date;
